It seems that I have discovered two major predictor logic bugs. (1). counTable didn't have double precission with scores. For some scores this was leading ton inf values, making counting scores impossible after a stage. (2) maxPredictionCount check was being done in a wrong phase. Now it is correcterd and according to specification

+++ simple_bs_predictor/subseqPredictor.h
@@ -16,7 +16,7 @@
 
 using namespace std;
 
-typedef map<int, float> countTable;
+typedef map<int, double> countTable;
 
 class subseqPredictor{
 public:
@@ -32,7 +32,7 @@ private:
 	countTable countTable;
 	bool stop;
 	int prediction;
-	float score;
+	double score;
 	void push(vector<int>, int, int, int);
 	void predict(int*, int, int, int, int);
     void generateSubqueries(int*, int);
